Split redirects api module into v2 and v3
