+++ wagtail/contrib/redirects/api/__init__.py
@@ -1,0 +1,7 @@
+from .v2 import RedirectsAPIViewSet, RedirectSerializer
+
+# For backwards compatibility
+__all__ = [
+    "RedirectSerializer",
+    "RedirectsAPIViewSet",
+]
