+++ wagtail/contrib/redirects/api/v3/__init__.py
@@ -1,0 +1,5 @@
+from .router import router
+
+__all__ = [
+    "router",
+]
